# src/public/swagger.yaml
openapi: 3.0.3
info:
  title: Tooth Fairy
  description: Tooth Fairy Rest APIs documentation
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/'
    description: Local Development server
...
paths:
...
  /quizz/{id}:
    delete:
      summary: create Quizz Content
      operationId:  Delete Quizz
      tags:
        - Quizz
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: ID of the quizz
          example: 62905e64f9a0fecd837e4ed6
      security:
        - ApiKeyAuth: []            
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '409':
          description: Conflict
        '500':
          description: Server Error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
...